Fix TypeError: Cannot read properties of null (reading '0')
Had to move the render fn into a new file
