+++ src/components/counter.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useState } from "react";
-// import { Card, CardContent, CardHeader, CardTitle } from "./ui/card";
-// import { Button } from "./ui/button";
+import { Card, CardContent, CardHeader, CardTitle } from "./ui/card";
+import { Button } from "./ui/button";
 
 export default function Counter({
   initialCount = 0,
@@ -11,14 +11,14 @@ export default function Counter({
   const [count, setCount] = useState(initialCount);
 
   return (
-    <div className="flex flex-col  items-center justify-center border-none shadow-none">
-      <div>
-        <div data-testid="current-count">{count}</div>
-      </div>
-      <div className="flex flex-col gap-2">
-        <button onClick={() => setCount(count + 1)}>Increment</button>
-        <button onClick={() => setCount(initialCount)}>Reset</button>
-      </div>
-    </div>
+    <Card className="flex flex-col  items-center justify-center border-none shadow-none">
+      <CardHeader>
+        <CardTitle data-testid="current-count">{count}</CardTitle>
+      </CardHeader>
+      <CardContent className="flex flex-col gap-2">
+        <Button onClick={() => setCount(count + 1)}>Increment</Button>
+        <Button onClick={() => setCount(initialCount)}>Reset</Button>
+      </CardContent>
+    </Card>
   );
 }
